Avoid preloading Faraday by using middleware API
4a4523aefe782f8d4eaeb6aa959173906bb83569 (v0.0.6) added support for the
middleware API (`Faraday::Middleware`), introduced in Faraday 0.8.0.

This change drops the support for the earlier style (which was effectively
dropped in later Faraday releases) to avoid loading Faraday
unnecessarily.

It brings the code in line with what we see on
https://github.com/lostisland/faraday_middleware/blob/e45e8f75a00e260eb7a4719f252883d3015b18d3/lib/faraday_middleware.rb#L26
which is a good indication that it's the way to go.

It was tested with Faraday 0.7.4 (just for sanity; it fails with that
version), then 0.8.0 and 1.0.1 (tests run fine with those),
therefore gemspec was bumped accordingly